apply MVVM
