+++ app/src/main/java/com/example/uhyo/view/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.uhyo.main
+package com.example.uhyo.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
