+++ app/src/main/java/com/example/uhyo/view/OnlineMenuActivity.kt
@@ -1,4 +1,4 @@
-package com.example.uhyo.online
+package com.example.uhyo.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
